% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemes.R
\name{applySchemeToTrack}
\alias{applySchemeToTrack}
\title{applySchemeToTrack}
\usage{
applySchemeToTrack(track, scheme = NULL)
}
\arguments{
\item{track}{the Gviz track to modify}

\item{scheme}{the name of the scheme to be applied as a string, or \code{NULL} (default) for the current scheme}
}
\value{
the track with modified display parameters
}
\description{
Applies a Gviz plotting scheme to the display parameters of a track
}
\author{
Fabian Mueller
}
