% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getGvm}
\alias{getGvm}
\title{getGvm}
\usage{
getGvm(assembly)
}
\arguments{
\item{assembly}{string specifying the genome assembly}
}
\value{
\code{GvizManager} object
}
\description{
retrieve the \code{GvizManager} object for a given genome assembly
}
