% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GvizManager-class.R
\docType{methods}
\name{getIdeogramTrack,GvizManager-method}
\alias{getIdeogramTrack,GvizManager-method}
\alias{getIdeogramTrack}
\title{getIdeogramTrack-methods}
\usage{
\S4method{getIdeogramTrack}{GvizManager}(object, chrom,
  applyScheme = TRUE)
}
\arguments{
\item{object}{\code{\linkS4class{GvizManager}} object}

\item{chrom}{name of the chromosome for which the track should be returned}

\item{applyScheme}{apply the display parameters of the currently active Gviz scheme to the result}
}
\value{
\code{Gviz::IdeogramTrack} for the specified chromosome
}
\description{
Get a \code{Gviz} ideogram track for a specified chromosome
}
\author{
Fabian Mueller
}
