% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gvm_hg19}
\alias{gvm_hg19}
\title{\code{Gviz} annotation data for geneome assembly hg19}
\format{A \code{GvizManager} object with annotation data for genome assembly
\describe{
  \item{gene annotation}{based on GENCODE v19}
}}
\usage{
gvm_hg19
}
\description{
\code{Gviz} annotation data for geneome assembly hg19
}
\keyword{datasets}
