% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gvm_hg38}
\alias{gvm_hg38}
\title{\code{Gviz} annotation data for geneome assembly hg38}
\format{A \code{GvizManager} object with annotation data for genome assembly
\describe{
  \item{gene annotation}{based on GENCODE v27}
}}
\usage{
gvm_hg38
}
\description{
\code{Gviz} annotation data for geneome assembly hg38
}
\keyword{datasets}
