% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gvm_mm10}
\alias{gvm_mm10}
\title{\code{Gviz} annotation data for geneome assembly mm10}
\format{A \code{GvizManager} object with annotation data for genome assembly
\describe{
  \item{gene annotation}{based on GENCODE M16}
}}
\usage{
gvm_mm10
}
\description{
\code{Gviz} annotation data for geneome assembly mm10
}
\keyword{datasets}
